select method, item_choice method and relevant tests moved to Menu
