+++ lib/menu.rb
@@ -1,7 +1,13 @@
 class Menu
 
-    def list
-        
+    def select
+        @item = item_choice 
+    end
+
+    DISHLIST = [:burger, :pizza, :chips]
+
+    def item_choice 
+        DISHLIST.sample
     end
 
 end
