+++ spec/takeaway_spec.rb
@@ -1,9 +1,7 @@
 require 'takeaway' 
 
 describe TakeAway do 
-    it 'responds to select method' do 
-        expect(subject).to respond_to(:select)
-    end
+
 
     # it 'checks that select method takes one argument' do
     #     expect(subject).to respond_to(:select).with(1).arguments
@@ -17,7 +15,5 @@ describe TakeAway do
         expect(subject.see_dishes).to eq(TakeAway::DISHLIST)
     end
 
-    it 'checks that select method returns a random item from DISHLIST' do 
-        expect(subject.select).to eq(:burger).or eq(:pizza).or eq(:chips)
-    end
+    
 end
